Typeorm relationships and Unknown Error handler

+++ server/src/v1/ideas/services/idea.service.ts
@@ -1,5 +1,6 @@
 import { HttpException, HttpStatus, Injectable, Logger } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
+import { UserEntity } from "src/v1/user/entity/User.entity";
 import { Repository } from "typeorm";
 import { CreateIdeaDto } from "../dtos/CreateIdea.dto";
 import { UpdateIdeaDtos } from "../dtos/UpdateIdea.dto";
@@ -7,27 +8,61 @@ import { IdeaEntity } from "../entity/idea.entity";
 
 @Injectable()
 export class IdeaServices {
-    private logger = new Logger('IdeaController');
-    constructor(@InjectRepository(IdeaEntity) private ideaRepository: Repository<IdeaEntity>) { }
+    
+    constructor(
+        @InjectRepository(IdeaEntity) private ideaRepository: Repository<IdeaEntity>,
+        @InjectRepository(UserEntity) private readonly userRepository:Repository<UserEntity>
+    ) { }
 
-    async getAllIdeas() {
-        return await this.ideaRepository.find();
+
+    toResponseObject(idea:IdeaEntity){
+        return {...idea, author: idea.author.toReponseObject(false) };
     }
 
-    async createIdea(createIdeaDto: CreateIdeaDto) {
-        this.logger.log(JSON.stringify(createIdeaDto));
-        return await this.ideaRepository.save(this.ideaRepository.create(createIdeaDto));
+    private ensureOwnerShip (idea:IdeaEntity, userId: number){
+        if(idea.author.id !== userId){
+            throw new HttpException('You don\'t belong here!',HttpStatus.UNAUTHORIZED);
+        }
+    }
+
+    async getAllIdeas() {
+        const ideas =  await this.ideaRepository.find({
+            relations: ['author']
+        });
+        return ideas.map(idea => this.toResponseObject(idea));
+    }
+
+    async createIdea(id: number,createIdeaDto: CreateIdeaDto) {
+        const user = await this.userRepository.findOne(id);
+        const idea =  await this.ideaRepository.save(
+            this.ideaRepository.create({
+                ...createIdeaDto,
+                author: user
+            })
+        );
+        return this.toResponseObject(idea);
+
     }
 
     async getOneIdea(id: number) {
-        const idea = await this.ideaRepository.findOne(id);
+        const idea = await this.ideaRepository.findOne({
+            where:{id},
+            relations: ['author']
+        });
         if (!idea) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
-        return idea;
+        return this.toResponseObject(idea);
     }
 
-    async updateIdea(id: number, updateIdeaDto: Partial<UpdateIdeaDtos>) {
-        this.logger.log(JSON.stringify(updateIdeaDto));
-        this.getOneIdea(id);
+    async updateIdea(
+        id: number,
+        userId: number,
+        updateIdeaDto: Partial<UpdateIdeaDtos>,
+        ) {
+        const idea = await this.ideaRepository.findOne(id, {
+            relations: ['author']
+        });
+        if(!idea) throw new HttpException('Idea not found!', HttpStatus.NOT_FOUND);
+        this.ensureOwnerShip(idea ,userId);
         try {
             await this.ideaRepository.update(id, updateIdeaDto);
         } catch (error) {
@@ -37,9 +72,16 @@ export class IdeaServices {
         return this.getOneIdea(id);
     }
 
-    async deleteOne(id: number) {
-        const idea = this.getOneIdea(id);
+    async deleteOne(id: number, userId: number) {
+        const idea = await this.ideaRepository.findOne(id,{
+            relations: ['author']
+        });
+        if(!idea) throw new HttpException('Idea not found!', HttpStatus.NOT_FOUND);
+        this.ensureOwnerShip(idea, userId);
         await this.ideaRepository.delete(id);
-        return idea;
+        return {
+            deleted: true,
+            ...idea.author.toReponseObject() 
+        };
     }
 }
